Implement ray and update PPM to gradient

+++ src/color.rs
@@ -1,6 +1,6 @@
 use crate::vec3::Color;
 
-const COLOR_MULTIPLIER: f32 = 255.99;
+const COLOR_MULTIPLIER: f64 = 255.99;
 
 pub fn write_color(pixel_color: Color) -> String {
     format!(
